fix: move captura de exceptions BadRequest e NotFound para o GlobalExceptionHandler correto
