+++ src/main/java/fly/be/flyflix/auth/exception/GlobalExceptionHandler.java
@@ -1,6 +1,9 @@
 package fly.be.flyflix.auth.exception;
 
 import fly.be.flyflix.auth.controller.dto.MensagemRespostaDTO;
+import fly.be.flyflix.conteudo.exceptions.BadRequestException;
+import fly.be.flyflix.conteudo.exceptions.DefaultMessageError;
+import fly.be.flyflix.conteudo.exceptions.NotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -40,5 +43,19 @@ public class GlobalExceptionHandler {
                 "ERRO_INTERNO"
         );
         return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(dto);
+    }
+
+    @ExceptionHandler(BadRequestException.class)
+    public ResponseEntity<DefaultMessageError> handlerBadRequestException(BadRequestException e) {
+        DefaultMessageError error = new DefaultMessageError(e.getStatusCode().value(), e.getReason());
+
+        return ResponseEntity.status(e.getStatusCode()).body(error);
+    }
+
+    @ExceptionHandler(NotFoundException.class)
+    public ResponseEntity<DefaultMessageError> handlerNotFoundException(NotFoundException e) {
+        DefaultMessageError error = new DefaultMessageError(e.getStatusCode().value(), e.getReason());
+
+        return ResponseEntity.status(e.getStatusCode()).body(error);
     }
 }
